starting to use Pathlib and new resultDir . Read all comments ...
Pathlib is nice, but it seems to break Snakemake-expands. That is for
now solved by casting that Path back to a string. It also means that
lot of old code is broken - and has been commented out for now.
This includes mod2 and mod5.

Also, results are grouped in resultPath