Pretty big rewrite. More under checkin-details
In order to force snakemake --report to output results in the same
order they are run, and in order to check results more 'at a glance'
several changes have been made.
There is now a file with meta-information about the rules, rules.yaml
Some of the stuff that used to be set directly in the pipeline is now
set there. This are things as detailed description of the rules, as
well as order the tests will be displayed in the --report.

It also has the potentional to be used to turn rules on/off at a later
stage. Test parameter could have been put in here, but right now these
still are found in config.yaml. In an ideal world, rules. yaml should
not need to be edited.

The pipeline so wark works, but plots need to be changed.