Gutorms sandbox (#1)
* New version Will Rayner perl script

Upgraded to 4.3 that supports a path to perl-script. Hence we do not
need to edit the script. It also seems faster and should use less
memory.

* Chromosome split in mod5

As well as some comments and editing/customizing of snakefile
Previous version of mod5 did not use results from Will Rayner

* typo/comments

* comment out split function

* comment out split/split_report/inputation

* minor tweaks to rerun pipeline

* diffs two plink directories with tweaks in .bim

* comments and test of dict_count_items

* rayner_report present

* Some more comments/rules ordering

pedigree_fix_pruned was (in rules file) in mod2
Better comments on m2_outpt

* did not work unless . was one of the directories

Also added set -x for debug (commented)

* Cleanup and more links

Note that we for now link to non-public sites ...

* comments formatted/updated

* playing with documentation

Testing out use of params and more references to
extra info (such as call rates)

* comment in egrep

* For parsing yaml.details files

* Warning when multiple matches found

* changed title and size of plot_point_and_line

* ignoring snpname when comparing

* change plotsize again.

* create_fam_map added + minor tweaks

This can now be used from other subdirs,
it won't crash (as it did).

* For creating 'published raw files'

* moved to the configs subdir

* removed debug output

* remove .nosex and general cleanup

* For Norment data retriaval 954

* Works for both norment 954 chips

Removed rule all for now. It is not needed

* more columns back when testing markerchanges

* new naming conventions

Both for files/directories, but also in the .fam file (fam id is
sentrix-id as well)

* destination renaming and v3 file

* now compares files and not directories

Stem is the imput. We don't need identical names anymore

* new output target

* rename of final rule (-> m1_output)

* Files moved to new aux/bin

* creates checksums

* R and python libraries

* environment moved to aux

* clearer docs and actual use of resultdir

* prior to allele flip hack (ignore strand order)

* checkUpdates can now ignore allele order

* untested but with all checkupdates .bim ignoring allele order

* allele_flip in checkupdate

* more robust alleleFlip

* bim files compared with alleles (also flipped)

Regression tests ok

* allele order is not significant when looking for match

* syntax error due to comment in continuation line. gah.

* Push before move to fhi-beta